+++ app/Http/Controllers/Admin/AdminEventoController.php
@@ -11,21 +11,25 @@ class AdminEventoController extends Controller
 {
     public function vista()
     {
+        $this->authorize('admin', User::class);
         return view('admin.eventos');
     }
 
     public function datos()
-    {    
+    {
+        $this->authorize('admin', User::class);
         return EventoLista::orderBy('created_at', 'desc')->paginate(20);
     }
 
     public function inscritos($evento)
-    {       
+    {
+        $this->authorize('admin', User::class);     
         return Evento::where('evento', $evento)->paginate(20);
     }
 
     public function crear(Request $request)
     {
+        $this->authorize('admin', User::class);
         $request->validate([ 'nombre' => 'required' ]);
 
         $evento = new EventoLista;            
@@ -36,6 +40,7 @@ class AdminEventoController extends Controller
 
     public function eliminar($id)
     {
+        $this->authorize('admin', User::class);
         EventoLista::find($id)->delete(); 
     }
 }
